+++ client/packages/lowcoder/src/comps/comps/chatComp/index.ts
@@ -1,0 +1,3 @@
+// client/packages/lowcoder/src/comps/comps/chatComp/index.ts
+export { ChatComp } from "./chatComp";
+export type { ChatCompProps } from "./chatCompTypes";
